Room uses more reliable scrolling (uses accel instead of pos)
