+++ projectiles/explosions.py
@@ -1,0 +1,54 @@
+import random as rand
+import time
+
+import pygame
+from pygame.math import Vector2 as vec
+
+import calculations as calc
+import classbases as cb
+import constants as cst
+
+
+class StdBulletExplode(cb.ActorBase):
+    def __init__(self, owner, pos_x: float, pos_y: float):
+        """An explosion that is displayed on-screen when a projectile hits something.
+
+        Args:
+            owner: The entity to render an explosion for
+            pos_x: The x-position to spawn the explosion at
+            pos_y: The y-position to spawn the explosion at
+        """
+        super().__init__()
+        self.layer = cst.LAYER['explosion']
+        self.show(self.layer)
+        self.owner = owner
+
+        self.pos = vec((pos_x, pos_y))
+        self.posOffset = vec(self.pos.x - self.owner.hit.pos.x, self.pos.y - self.owner.hit.pos.y)
+        self.set_images('sprites/bullets/bullets.png', 32, 32, 8, 4, 0, 1)
+        self.set_rects(self.pos.x, self.pos.y, 32, 32, 32, 32)
+
+        self.render_images()
+        self.randRotation = rand.randint(1, 360)
+
+    def movement(self):
+        self.center_rects()
+        self.pos = self.owner.hit.hitbox.center + self.posOffset
+
+    def update(self):
+        self.movement()
+
+        if calc.get_time_diff(self.last_frame) >= cst.SPF:
+            if self.index > 3:
+                self.kill()
+            else:
+                self.render_images()
+                self.index += 1
+            self.last_frame = time.time()
+
+        # Give explosion its random rotation
+        self.image = pygame.transform.rotate(self.orig_image, self.randRotation)
+        self.center_rects()
+
+    def __repr__(self):
+        return f'ProjExplode({self.owner}, {self.pos})'
